client/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
ament-clang-format==0.12.14
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
anyio==4.12.1
apriltag-msgs==2.0.1
aruco-msgs==5.0.5
aruco-opencv-msgs==2.3.1
attrs==25.4.0
blinker==1.9.0
bond==4.1.2
builtin-interfaces==1.2.2
CacheControl==0.14.4
camera-calibration==3.0.8
cartographer-ros-msgs==2.0.9002
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cobot1==0.0.0
coloredlogs==15.0.1
composition-interfaces==1.2.2
control-msgs==4.8.0
controller-manager==2.52.2
controller-manager-msgs==2.52.2
cryptography==46.0.5
cv-bridge==3.2.1
demo-nodes-py==0.20.7
depthai-ros-driver==2.11.2
depthai-ros-msgs==2.11.2
diagnostic-aggregator==4.0.6
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
dsr-bringup2==0.1.2
dsr-example==0.0.0
dsr-msgs2==1.1.0
dsr-mujoco==0.1.0
dsr-tests==0.0.0
dsr-visualservoing==0.0.0
dwb-msgs==1.1.20
dynamixel-sdk==3.8.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
exceptiongroup==1.3.1
ffmpeg-image-transport-msgs==1.1.2
firebase_admin==7.1.0
Flask==3.1.3
flatbuffers==25.12.19
foxglove-msgs==3.2.2
gazebo-model-attachment-plugin==1.0.3
gazebo-model-attachment-plugin-msgs==1.0.3
gazebo-msgs==3.9.0
gazebo-video-monitor-interfaces==0.8.1
gazebo-video-monitor-utils==0.8.1
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gps-msgs==2.1.1
grid-map-msgs==2.0.1
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
interactive-markers==2.3.2
irobot-create-common-bringup==2.1.0
irobot-create-msgs==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
laser-geometry==2.4.1
launch==1.0.13
launch-param-builder==0.1.1
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
lazy_loader==0.4
lifecycle-msgs==1.2.2
llvmlite==0.46.0
logging-demo==0.20.7
map-msgs==2.1.0
MarkupSafe==3.0.3
message-filters==4.3.11
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
msgpack==1.1.2
nav-2d-msgs==1.1.20
nav-msgs==4.9.0
nav2-common==1.1.20
nav2-msgs==1.1.20
nav2-simple-commander==1.0.0
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
onnxruntime==1.23.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
osrf-pycommon==2.1.6
packaging==26.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
PyMatting==1.1.15
python-qt-binding==1.1.3
qrcode==8.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.7
realsense2-camera-msgs==4.56.4
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
robot-upstart==1.0.4
ros-gz-interfaces==0.244.20
ros-ign-interfaces==0.244.20
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2controlcli==2.52.2
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.30.0
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-monitor==1.0.6
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
rtabmap-msgs==0.22.1
rtabmap-python==0.1.0
scikit-image==0.25.2
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
simple-term-menu-vendor==1.5.7
slam-toolbox==2.6.10
smclib==4.1.2
srdfdom==2.0.7
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.17
tf2-kdl==0.25.17
tf2-msgs==0.25.17
tf2-py==0.25.17
tf2-ros-py==0.25.17
tf2-sensor-msgs==0.25.17
tf2-tools==0.25.17
theora-image-transport==2.5.4
tifffile==2025.5.10
topic-monitor==0.20.7
tqdm==4.67.3
trajectory-msgs==4.9.0
turtle-tf2-py==0.3.7
turtlebot3-applications-msgs==1.0.1
turtlebot3-aruco-tracker==1.3.3
turtlebot3-automatic-parking==1.3.3
turtlebot3-automatic-parking-vision==1.3.3
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.3.3
turtlebot3-home-service-challenge-aruco==1.0.5
turtlebot3-home-service-challenge-core==1.0.5
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.3
turtlebot3-yolo-object-detection==1.3.3
turtlebot4-diagnostics==1.0.3
turtlebot4-msgs==1.0.5
turtlebot4-navigation==1.0.5
turtlebot4-python-tutorials==1.0.1
turtlebot4-setup==1.0.6
turtlebot4-tests==1.0.3
turtlesim==1.4.3
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.6.3
velodyne-msgs==2.5.1
vision-msgs==4.1.1
visualization-msgs==4.9.0
Werkzeug==3.1.6
xacro==2.1.1
